sound: Merge dorimanx fixes for Faux Sound 1/2
This is an adaptation from his fixes, all credits go to: @dorimanx

https://github.com/dorimanx/DORIMANX_LG_STOCK_LP_KERNEL/commit/298f96bcb82347c22101ea1da31b5beaf5e5d361
https://github.com/dorimanx/DORIMANX_LG_STOCK_LP_KERNEL/commit/8c1a50970f74cd76cd2ba4abbce17506145a99a7
https://github.com/dorimanx/DORIMANX_LG_STOCK_LP_KERNEL/commit/d16156ea71964cb5f14c266b0f9af9326d67cd40
https://github.com/dorimanx/DORIMANX_LG_STOCK_LP_KERNEL/commit/06e3a450a90fab775e1b817da750b73da2db253f

Also bumped the version, I have no idea in which version is my upstream
(Faux) but as long as it works as intended it's fine to use the
versioning to define patches.
Might need more fixes since I've picked 3.7 version of this driver and
it has a little more code than dori's one in page 118 but a quick reviewing should
be enough with some testing.
Edit: Rolling some pages forward I see there are more patches and fixes
to import, so... let's call this commit the 1/2.

Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>